+++ src/components/Employee/Deductions/DeductionsComponents.tsx
@@ -1,8 +1,6 @@
 import type { Garnishment } from '@gusto/embedded-api/models/components/garnishment'
-import { IncludeDeductionsForm } from './IncludeDeductionsForm/IncludeDeductionsForm'
 import { EmptyState } from './EmptyState/EmptyState'
 import { DeductionsList } from './DeductionsList/DeductionsList'
-import { DeductionsForm } from './DeductionsForm/DeductionsForm'
 import { DeductionsFormV2 } from './DeductionsForm/DeductionsFormV2'
 import { useFlow, type FlowContextInterface } from '@/components/Flow/useFlow'
 import type { componentEvents } from '@/shared/constants'
